requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
websockets>=11.0
termcolor>=2.0.0
rich>=13.0.0
lightgbm>=4.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
